// packages/astro-loader-github-prs/src/graphql/types.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A (potentially binary) string encoded using base64. */
  Base64String: { input: any; output: any }
  /**
   * Represents non-fractional signed whole numeric values. Since the value may
   * exceed the size of a 32-bit integer, it's encoded as a string.
   */
  BigInt: { input: any; output: any }
  /** An ISO-8601 encoded date string. */
  Date: { input: any; output: any }
  /** An ISO-8601 encoded UTC date string. */
  DateTime: { input: any; output: any }
  /** A Git object ID. */
  GitObjectID: { input: any; output: any }
  /** A fully qualified reference name (e.g. `refs/heads/master`). */
  GitRefname: { input: any; output: any }
  /** Git SSH string */
  GitSSHRemote: { input: any; output: any }
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: { input: any; output: any }
  /** A string containing HTML code. */
  HTML: { input: any; output: any }
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: { input: any; output: any }
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: { input: any; output: any }
  /** A valid x509 certificate string */
  X509Certificate: { input: any; output: any }
}
// ...
export enum PullRequestState {
  /** A pull request that has been closed without being merged. */
  Closed = 'CLOSED',
  /** A pull request that has been closed by being merged. */
  Merged = 'MERGED',
  /** A pull request that is still open. */
  Open = 'OPEN',
}
// ...
export type GetPrsQueryVariables = Exact<{
  search: Scalars['String']['input']
  first: Scalars['Int']['input']
  cursor?: InputMaybe<Scalars['String']['input']>
}>

export type GetPrsQuery = {
  search: {
    nodes?: Array<{
      id: string
      url: any
      title: string
      titleHTML: any
      number: number
      state: PullRequestState
      isDraft: boolean
      body: string
      bodyHTML: any
      bodyText: string
      createdAt: any
      mergedAt?: any | null
      author?:
        | { login: string; url: any; avatarUrl: any }
        | { login: string; url: any; avatarUrl: any }
        | { login: string; url: any; avatarUrl: any }
        | { login: string; url: any; avatarUrl: any }
        | { login: string; url: any; avatarUrl: any }
        | null
      repository: {
        name: string
        nameWithOwner: string
        url: any
        stargazerCount: number
        isInOrganization: boolean
        owner:
          | { login: string; url: any; avatarUrl: any }
          | { login: string; url: any; avatarUrl: any }
      }
    } | null> | null
    pageInfo: { hasNextPage: boolean; endCursor?: string | null }
  }
}